Fix GUI to build with Windows (#2153)
* Fix GUI to build with Windows

* Install Open3DViewer for Windows (and any other platfom)

* Add the actual files, which somehow didn't make it in the previous commit

* Fixed Window::PostRedraw() not redrawing on Windows

* Fix Python hanging after gui.application.run() finishes on Windows

* Style fixes, fix incomplete merge conflict with pybinds

* Fixed warning->error on Windows

* debugging fix for some windows options

* test CI fix for RelWithDebInfo

* Nope, RelWithDebInfo wants release libraries, not debug

* ON, ON, Debug -> mtd

* Try a generator expression instead

* Copy/paste error

* Fix build errors resulting from pulling from master

* Fixed test link/build, removed commented and unnecessary code from cmake

* Reduce disk space usage in win32 debug builds

* Added debugging to Windows CI

* Maybe GitHub Actions didn't like the step I added

* Fix yaml syntax

* Try Get-FsrmQuota

* Doesn't seem like it supports quotas, try just disk space

* Update for newer PowerShell command

* Fixes for style

* Remove unnecessary debug logging

+++ cpp/open3d/visualization/rendering/Light.h
@@ -47,8 +47,8 @@ struct Light {
     Eigen::Vector3f direction = Eigen::Vector3f(0.f, 0.f, -1.f);
 
     // Spot lights parameters
-    float light_cone_inner = M_PI / 4.f;
-    float light_cone_outer = M_PI / 2.f;
+    float light_cone_inner = float(M_PI / 4.0);
+    float light_cone_outer = float(M_PI / 2.0);
 };
 
 }  // namespace rendering
